+++ launcher/main/applications.h
@@ -31,6 +31,7 @@ typedef struct retro_app_s
     } paths;
     size_t crc_offset;
     retro_file_t *files;
+    size_t files_capacity;
     size_t files_count;
     bool use_crc_covers;
     bool crc_scan_done;